pyproject.toml

[project]
name = "enocean-ble"
version = "0.1.0"
description = ""
authors = [
    {name = "Christian Ohde", email = "christian.s1989@gmail.com"},
]
dependencies = [
    "bleak>=0.20.2",
    "typer>=0.9.0",
    "bluetooth-sensor-state-data>=1.6.1",
    "bluetooth-data-tools>=1.2.0",
    "home-assistant-bluetooth>=1.10.0",
    "sensor-state-data>=2.16.0",
    "pycryptodome>=3.18.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
    "ipython>=8.14.0",
    "pytest>=7.4.0",
]
